backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.27
alembic>=1.13.1
psycopg2-binary>=2.9.9
asyncpg>=0.29.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9
pydantic[email]>=2.6.2
python-dotenv>=1.0.1
weasyprint>=61.0
pytest>=8.0.2
httpx>=0.27.0
